refactor(graphql): split element schema modules
